added strictnullchecks policy * fixed all issues that were breaking * added custom errors for wc, turn14, and amqpPro

+++ src/productMgmt/jobs/brandActivationSequence.ts
@@ -1,9 +1,9 @@
-import { Inject, Service } from 'typedi';
-import { ProductMgmtService } from '../services/productMgmtService';
-import { ActiveBrandDTO } from '../dtos/activeBrandDto';
-import { ApiUserService } from '../../apiUsers/services/apiUserService';
 import _ from 'lodash';
+import { Inject, Service } from 'typedi';
+import { ApiUserService } from '../../apiUsers/services/apiUserService';
+import { ActiveBrandDTO } from '../dtos/activeBrandDto';
 import { PmgmtDTO } from '../dtos/pmgmtDto';
+import { ProductMgmtService } from '../services/productMgmtService';
 
 /**
  * BrandActivationSequence.
@@ -28,30 +28,33 @@ export class BrandActivationSequence {
     console.info('🔨 Import Brands Sequence Job starting!');
 
     const apiUser = await this.apiUserService.retrieve(activeBrandDto.userId);
-    const brandIds = apiUser.brandIds;
 
-    const pmgmtDto = new PmgmtDTO(
-      apiUser.siteUrl,
-      apiUser.turn14Keys,
-      apiUser.wcKeys,
-      activeBrandDto.brandId
-    );
+    if (apiUser != null) {
+      const brandIds = apiUser.brandIds;
 
-    if (activeBrandDto.active) {
-      if (!brandIds.includes(activeBrandDto.brandId)) {
-        brandIds.push(activeBrandDto.brandId);
-        apiUser.brandIds = brandIds;
-        await this.apiUserService.update(apiUser.id, apiUser);
+      const pmgmtDto = new PmgmtDTO(
+        apiUser.siteUrl,
+        apiUser.turn14Keys,
+        apiUser.wcKeys,
+        activeBrandDto.brandId
+      );
 
-        this.productMgmtService.import(pmgmtDto);
-      }
-    } else {
-      if (brandIds.includes(activeBrandDto.brandId)) {
-        _.pull(brandIds, activeBrandDto.brandId);
-        apiUser.brandIds = brandIds;
-        await this.apiUserService.update(apiUser.id, apiUser);
+      if (activeBrandDto.active) {
+        if (!brandIds.includes(activeBrandDto.brandId)) {
+          brandIds.push(activeBrandDto.brandId);
+          apiUser.brandIds = brandIds;
+          await this.apiUserService.update(apiUser.id, apiUser);
 
-        this.productMgmtService.delete(pmgmtDto);
+          this.productMgmtService.import(pmgmtDto);
+        }
+      } else {
+        if (brandIds.includes(activeBrandDto.brandId)) {
+          _.pull(brandIds, activeBrandDto.brandId);
+          apiUser.brandIds = brandIds;
+          await this.apiUserService.update(apiUser.id, apiUser);
+
+          this.productMgmtService.delete(pmgmtDto);
+        }
       }
     }
   }
